modify table column, index, trigger
