+++ com.hangum.tadpole.application/plugin.xml
@@ -28,7 +28,7 @@
             favicon="icons/TadpoleDBHub.ico"
             id="com.hangum.tadpole.browser.rap.branding"
             themeId="org.eclipse.rap.design.example.business.theme"
-            title="Tadpole DB Hub-v1.7.1(r1)">
+            title="Tadpole DB Hub-v1.7.1(r2)">
       </branding>
    </extension>
    <extension
