+++ com.hangum.tadpole.commons.libs/src/com/hangum/tadpole/commons/libs/core/define/SystemDefine.java
@@ -22,7 +22,7 @@ public class SystemDefine {
 	
 	public static final String NAME = "Tadpole DB Hub";
 	public static final String MAJOR_VERSION = "1.7.1";
-	public static final String SUB_VERSION = "Build(r1)";
+	public static final String SUB_VERSION = "Build(r2)";
 	public static final String RELEASE_DATE = "2016.04.13";
 	public static final String INFORMATION = "http://hangum.github.io/TadpoleForDBTools/";
