+++ com.hangum.tadpole.rdb.core/bin/com/hangum/tadpole/rdb/core/messages.properties
@@ -646,7 +646,7 @@ ResultSetComposite_16 = Column detail viewer
 
 ResultSetComposite_2 = Column to editor
 
-ResultSetComposite_3 = Request status
+ResultSetComposite_3 = Request query
 
 ResultSetComposite_6 = Select row
