+++ com.hangum.tadpole.rdb.core/bin/com/hangum/tadpole/rdb/core/messages_ko.properties
@@ -713,7 +713,7 @@ ResultSetComposite_16 = \uCEEC\uB7FC \uC0C1\uC138 \uBCF4\uAE30
 
 ResultSetComposite_2 = \uACB0\uACFC\uB97C \uC5D0\uB514\uD130\uB85C
 
-ResultSetComposite_3 = \uCFFC\uB9AC\uC694\uCCAD \uC0C1\uD0DC
+ResultSetComposite_3 = \uCFFC\uB9AC\uC694\uCCAD
 
 ResultSetComposite_6 = \uC5F4\uC744 \uC120\uD0DD\uD558\uC138\uC694.
