+++ com.hangum.tadpole.rdb.core/src/com/hangum/tadpole/rdb/core/viewers/object/sub/rdb/table/trigger/TriggerViewFilter.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     hangum - initial API and implementation
  ******************************************************************************/
-package com.hangum.tadpole.rdb.core.viewers.object.sub.rdb.trigger;
+package com.hangum.tadpole.rdb.core.viewers.object.sub.rdb.table.trigger;
 
 import org.eclipse.jface.viewers.Viewer;
